requirements.txt
requests==2.30.0
python-dotenv==1.0.0
python-telegram-bot==13
